requirements.txt
asgiref==3.4.1
Django==3.2.5
django-ckeditor==6.1.0
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
django-mptt==0.12.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
news==1.0
Pillow==8.3.1
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
